% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/aaa.R
\name{createRadShiny}
\alias{createRadShiny}
\alias{multipageshiny}
\title{Create multi-page shiny app}
\usage{
createRadShiny(dir = getwd(), nav, packages = .additionalPackages)
}
\arguments{
\item{dir}{directory under which app structure will be built. Defaults to cwd}

\item{nav}{A list or list of lists depicting the names in the navigation menu
  and submenus.}
}
\value{
Returns TRUE if everything worked out okay, else FALSE.
}
\description{
Create multi-page shiny app in the specified directory with optional
navigation list.
}
\details{
Lengthy description of this simple function
}
\examples{
createRadShiny(nav = list('home', 'p1', 'p2' = list('p2.1', 'p2.2')))
}
\seealso{
\href{https://github.com/vnijs/radiant}{Radiant}
}

