% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/aaa.R
\docType{package}
\name{radshiny}
\alias{radshiny}
\alias{radshiny-package}
\title{radshiny: A simple package to create multipage shiny apps}
\description{
The package provides one function, createRadShiny, which creates a shiny app in
the specified directory based on the navigation described in the input.
}
\section{Doesn't do much...}{

}

